refactor: CamelCase to LowerCase
