--- evertecine/lib/data/movie/IMovieRepository.dart
@@ -1,9 +1,0 @@
-
-import 'package:evertecine/network/core/BaseNetworkResponse.dart';
-
-import '../../domain/model/Movie.dart';
-
-abstract class IMovieRepository {
-  Future<BaseNetworkResponse<Movie>> getUpcomingMovies({int page = 1});
-}
-
