+++ evertecine/lib/data/movie/i_movie_repository.dart
@@ -1,0 +1,9 @@
+
+import 'package:evertecine/network/core/base_network_response.dart';
+
+import '../../domain/model/movie.dart';
+
+abstract class IMovieRepository {
+  Future<BaseAPIResponse<Movie>> getUpcomingMovies({int page = 1});
+}
+
