+++ evertecine/lib/data/movie/network/network_movies_repository_impl.dart
@@ -1,10 +1,10 @@
 import 'dart:convert';
-import 'package:evertecine/data/movie/network/INetworkMovieRepository.dart';
-import 'package:evertecine/domain/model/Movie.dart';
-import 'package:evertecine/data/movie/IMovieRepository.dart';
+import 'package:evertecine/data/movie/network/i_network_movie_repository.dart';
+import 'package:evertecine/domain/model/movie.dart';
+import 'package:evertecine/data/movie/i_movie_repository.dart';
 import 'package:http/http.dart' as http;
 
-import '../../../network/core/BaseNetworkResponse.dart';
+import '../../../network/core/base_network_response.dart';
 
 class NetworkMoviesRepositoryImpl implements Inetworkmovierepository {
   final String _baseUrl = 'https://api.themoviedb.org/3';
@@ -17,23 +17,23 @@ class NetworkMoviesRepositoryImpl implements Inetworkmovierepository {
   }
 
   @override
-  Future<BaseNetworkResponse<Movie>> getUpcomingMovies({int page = 1}) async {
+  Future<BaseAPIResponse<Movie>> getUpcomingMovies({int page = 1}) async {
     final url = Uri.parse('$_baseUrl/movie/upcoming?api_key=$_apiKey&language=pt-BR&page=$page');
 
     try {
       final response = await http.get(url);
 
       if (response.statusCode != 200) {
-        return BaseNetworkResponse(success: false, message: "Request failed", statusCode: response.statusCode);
+        return BaseAPIResponse(success: false, message: "Request failed", statusCode: response.statusCode);
       }
       final Map<String, dynamic> data = json.decode(response.body);
       final List<dynamic> results = data['results'];
 
       final List<Movie> movies = results.map((item) => Movie.fromNetwork(item, _genres)).toList();
-      if (movies.isEmpty) return BaseNetworkResponse(success: false, message: "Empty movies list");
-      return BaseNetworkResponse(success: true, results: movies, statusCode: 200);
+      if (movies.isEmpty) return BaseAPIResponse(success: false, message: "Empty movies list");
+      return BaseAPIResponse(success: true, results: movies, statusCode: 200);
     } catch (e) {
-      return BaseNetworkResponse(success: false, message: e.toString());
+      return BaseAPIResponse(success: false, message: e.toString());
     }
   }
 
@@ -57,22 +57,22 @@ class NetworkMoviesRepositoryImpl implements Inetworkmovierepository {
   }
 
   @override
-  Future<BaseNetworkResponse<Movie>> searchMovieByName(String value) async {
+  Future<BaseAPIResponse<Movie>> searchMovieByName(String value) async {
     final url = Uri.parse('$_baseUrl/search/movie?api_key=$_apiKey&query=$value&language=pt-BR');
     try {
       final response = await http.get(url);
 
       if (response.statusCode != 200) {
-        return BaseNetworkResponse(success: false, message: "Request failed", statusCode: response.statusCode);
+        return BaseAPIResponse(success: false, message: "Request failed", statusCode: response.statusCode);
       }
       final Map<String, dynamic> data = json.decode(response.body);
       final List<dynamic> results = data['results'];
 
       final List<Movie> movies = results.map((item) => Movie.fromNetwork(item, _genres)).toList();
-      if (movies.isEmpty) return BaseNetworkResponse(success: false, message: "Empty movies list");
-      return BaseNetworkResponse(success: true, results: movies, statusCode: 200);
+      if (movies.isEmpty) return BaseAPIResponse(success: false, message: "Empty movies list");
+      return BaseAPIResponse(success: true, results: movies, statusCode: 200);
     } catch (e) {
-      return BaseNetworkResponse(success: false, message: e.toString());
+      return BaseAPIResponse(success: false, message: e.toString());
     }
 
   }
