+++ evertecine/lib/domain/model/movie.dart
