+++ evertecine/lib/network/core/base_network_response.dart
@@ -1,0 +1,15 @@
+class BaseAPIResponse<T> {
+  final bool success;
+  final String? message;
+  final int? statusCode;
+  final int? page;
+  final List<T>? results;
+
+  BaseAPIResponse({
+    required this.success,
+    this.message,
+    this.statusCode,
+    this.page,
+    this.results,
+  });
+}
