+++ evertecine/lib/ui/components/movie_card_item.dart
@@ -2,7 +2,7 @@ import 'package:evertecine/ui/components/styles.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import '../../domain/model/Movie.dart';
+import '../../domain/model/movie.dart';
 
 class MovieCardItem extends StatelessWidget {
   final Movie movie;
